requirements.txt
aiohttp       == 3.6.2
async-timeout == 3.0.1
attrs         == 20.2.0
chardet       == 3.0.4
discord.py    == 1.5.0
idna          == 2.10
multidict     == 4.7.6
pip           == 20.1.1
setuptools    == 50.3.0
wheel         == 0.35.1
yarl          == 1.6.0
python-dotenv == 0.14.0